chore: Big cleanup of the files
