+++ packages/article-main-comment/src/article-label/article-label.js
@@ -2,7 +2,7 @@ import React from "react";
 import { View } from "react-native";
 import ArticleLabel from "@times-components/article-label";
 import Context from "@times-components/context";
-import styles from "../styles/article-header";
+import styles from "../styles";
 
 export default (props) => (
   <Context.Consumer>
